added debugger mode option in quiz
